+++ installer/api.py
@@ -59,13 +59,19 @@ def sendApiPOST(retCode: int=0):
 		url = API_URL.split("/", maxsplit=1)
 		path = f"/{url[1]}"
 		url = url[0]
-		conn = http.client.HTTPSConnection(url, context=ssl._create_unverified_context()) # Unverified context because url does not have an ssl certificate
+		conn = http.client.HTTPSConnection(url, timeout=2, context=ssl._create_unverified_context()) # Unverified context because url does not have an ssl certificate
 
-		# Send the POST request
-		conn.request("POST", path, params, headers)
-
-		# Close the connection
-		conn.close()
+		try:
+			# Send the POST request
+			conn.request("POST", path, params, headers)
+	
+			# Get response from server
+			conn.getresponse()
+	
+			# Close the connection
+			conn.close()
+		except Exception as e:
+			pass
 	
 ####################### UTILS FUNCTIONS #######################
 def getOSReleaseName() -> str: